+++ js/photo.js
@@ -1,0 +1,29 @@
+const PHOTOS = [
+  "пляж",
+  "указатель",
+  "бухта",
+  "девушка",
+  "суп",
+  "машина",
+  "клубника",
+  "сок",
+  "самолет",
+  "обувь",
+  "пляж",
+  "ауди",
+  "сельдерей",
+  "котенок",
+  "космо-тапки",
+  "облака",
+  "хор",
+  "коломбо",
+  "тапки",
+  "пальмы",
+  "курица",
+  "закат",
+  "краб",
+  "концерт",
+  "бегемот",
+];
+
+export {PHOTOS};
