+++ js/stringLength.js
@@ -1,0 +1,4 @@
+const getStringLength = function  (valueString, maxLength) {
+  return valueString.length <= maxLength;
+}
+export {getStringLength};
